+++ cognite/powerops/resync/models/market/base.py
@@ -4,7 +4,7 @@ from typing import ClassVar, Union
 
 from pydantic import Field
 
-from cognite.powerops.resync.config_classes.cdf_labels import AssetLabel
+from cognite.powerops.cdf_labels import AssetLabel
 from cognite.powerops.resync.models.base import AssetType, NonAssetType
