+++ cognite/powerops/resync/models/market/benchmark.py
@@ -5,7 +5,7 @@ from typing import ClassVar
 from cognite.client.data_classes import TimeSeries
 from pydantic import Field
 
-from cognite.powerops.resync.config_classes.cdf_labels import AssetLabel
+from cognite.powerops.cdf_labels import AssetLabel
 from cognite.powerops.resync.models.base import NonAssetType
 from cognite.powerops.resync.models.market.base import Bid, Process
